+++ tests/integration-network/src/mockchain/mod.rs
@@ -1,5 +1,6 @@
 //! Integration tests which exercise contract deployment and execution on a mock chain.
 
+mod asset_amount;
 mod counter;
 mod fpi;
 mod notes;